+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add option of multiple `VectorIndex` implementations
 - Batch create tables, put-components
 - Add fix for exporting component with `CFG.json_native=True`
+- Add compute_kwargs to `Component` and `Job`
 
 #### New Features & Functionality